Remove state footgun in evaluator design
**All** state that must persist should be in the 'state'
parameter, rather than in instance state. Only one instance of
the evaluator is allowed per loop (so they can share api things like
the docker client). Split methods out so data (which can be stored
in db) and compute are treated very differently.

This is actually a mess to do for localprocess so that is completely
removed.